+++ src/database/migrations/2020-11-23-195700_create_products/up.sql
@@ -4,11 +4,11 @@ CREATE TABLE products (
   name VARCHAR NOT NULL,
   description VARCHAR NOT NULL,
   store_name VARCHAR,
-  store_price DECIMAL,
-  price DECIMAL NOT NULL,
-  additional_charge DECIMAL DEFAULT 0.00,
+  store_price REAL,
+  price REAL NOT NULL,
+  additional_charge REAL DEFAULT 0.00,
   color VARCHAR NOT NULL,
-  weight DECIMAL NOT NULL,
+  weight REAL NOT NULL,
   brand VARCHAR NOT NULL,
   stock_amount INTEGER NOT NULL
 );
